LayoutTests/imported/w3c/web-platform-tests/webrtc-extensions/RTCRtpReceiver-jitterBufferTarget-stats-expected.txt


FAIL measure raising and lowering video jitterBufferTarget assert_equals: jitterBufferTarget supported for video expected (object) null but got (undefined) undefined
FAIL measure raising and lowering audio jitterBufferTarget assert_equals: jitterBufferTarget supported for audio expected (object) null but got (undefined) undefined
